[dev_model/infra] recovery infra_timer and replace HAL_Random with HAL_UptimeMs

+++ src/infra/infra_timer.h
@@ -17,17 +17,17 @@ typedef struct {
 
 void iotx_time_start(iotx_time_t *timer);
 
-uint32_t utils_time_spend(iotx_time_t *start);
+uint64_t utils_time_spend(iotx_time_t *start);
 
-uint32_t iotx_time_left(iotx_time_t *end);
+uint64_t iotx_time_left(iotx_time_t *end);
 
-uint32_t utils_time_is_expired(iotx_time_t *timer);
+int utils_time_is_expired(iotx_time_t *timer);
 
 void iotx_time_init(iotx_time_t *timer);
 
-void utils_time_countdown_ms(iotx_time_t *timer, uint32_t millisecond);
+void utils_time_countdown_ms(iotx_time_t *timer, uint64_t millisecond);
 
-uint32_t utils_time_get_ms(void);
+uint64_t utils_time_get_ms(void);
 
 #endif /* _IOTX_COMMON_TIMER_H_ */
